Simplify clue positioning
On mobiles, clues are shown on the top or bottom. On small screens this 
means you cannot always see the clue when the keyboard is open.

Also, removing the `isIOS` check so that we always position using the 
same mechanism (rather than using fixed) to simiplify the number of 
cases.