Fix entry highlighting
This relied on the main.js file which was no longer being used. I've 
removed that tried to prevent double-focusing- clicking again is 
currently broken (on guardian.com too) and causes focus to be lost. I 
think it's simplier to instead always focus the start of the entry when 
a clue is clicked.