Simplify & update eslint config
I found the linting hard to work with because it wasn't what I was used 
to and the config was split across two files. 

I've chosen the airbnb config (which it was already using) and removed 
prettier. I've disabled all failing rules that couldn't be autofixed- 
these could be looked individually.

I also find the condensed spacing much easier to work with. Perhaps 
everyone at the Guardian has wider displays than me.