Allow numbers to be inputted into crosswords
Sometimes crossword answers have numbers in them. Recently 2 down & 12
across in Weekend crossword No 410[1] are examples of this.

The existing regular expression to validate input doesn't allow digits,
causing nothing to happen when they are inputted.

I considered removing this check and replacing it with somthing simpler-
such as any single non-whitespace character but I'm not quite sure why
it exists at the moment and it has recently been extended in this way to
allow accented characters.

[1] https://www.theguardian.com/crosswords/weekend/410