Allow clues containing numbers or accents to be marked as answered
This regex is used to determine if the clue should be marked as answered
and so greyed out in the clue list. Previously the check that it did was
more restrictive than the check on input which meant an answer that
contained accents would never cause a clue to be marked as answered. In
the previous commit I also allowed answers to contain numbers which
would have a similar problem without this change.

I could have fixed this by making this regex the same as the input
validation in crossword.js by duplicating it or sharing the code.
However, I think a simpler thing to do is to make this less restrictive-
if we are validating characters on input all we need to do here is check
that a character has been inputted by making sure the value isn't the
empty string. The code also works if the value is `null` or `undefined`.