Remove broken error highlighting
I belive this is meant to cause the corrected cell to breifly flash but 
this doesn't work (at least on Chrome/Webkit browsers). Remove it, as I 
think it adds more complexity to crossword.js. If we wanted this 
behavior there might be better ways to do it (such as with CSS 
transitions).