Allow external components to set the grid state
This is needed so that multiplayer-crosswords can update the whole grid 
if the re-sync with the server if the connection is lost.

I think the react-y thing to do in this case would be to hoist the state 
up and re-arch the component to rely on props instead however, this 
would mean the component wouldn't be as easily re-usable since the 
wrapping component would be needed.