Don't reset anagrams when someone else is typing
When someone else is typing, the crossword component re-renders which 
triggers this function that causes the anagram letters to reset. 
Something is needed here for the case that you navigate to the next clue 
by the keyboard or click on another clue while the anagram helper is 
open. Instead set a key which causes a new anagram component to render 
and the state to reset. The newer react docs suggest this for this case 
rather than the unsafe `componentWillReceiveprops` function[1]

[1] 
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#recommendation-fully-uncontrolled-component-with-a-key