Fix positioning of '-' in anagram helper.
Previously the dash jumped around depending on if there was a value in 
the cell that was before it. I've switched it to use absolute 
positioning so it's positioned based on the cells edge. I've also 
tweaked the font size and z-index so it looks better relative to the 
text and is drawn above the border of the next cell.