Don't reset anagram when component re-renders
I thought I had fixed this issue in [1] but it was still a problem. It 
looks like it is happening because when moves are received it causes the 
anagram helper to re-render, and the shuffle method is called within the 
render.

Instead store the shuffled letters as state so they don't change when 
the component is re-rendered.

[1] 57f5bd399b6c526407f9180df2b67648b79c7fb4